ai-rag-knowledge-assistant/requirements.txt
streamlit>=1.33
python-dotenv>=1.0
faiss-cpu>=1.8.0
numpy>=1.26
openai>=1.30
